feat(testing): add combat.testing.ts helpers
 - remove duplicated helper code from combat tests
 - prefix all helpers with testCombat[HelperName] so they don't accidentally get used in the app
